Backend_v2/requirements.txt
# Web 框架
quart>=0.20.0,<0.21.0
quart-cors>=0.8.0,<0.9.0
quart-jwt-extended>=0.1.0,<0.2.0
werkzeug>=3.1.3,<4.0.0

# 异步支持
aiohttp>=3.11.18,<4.0.0
asyncpg>=0.30.0,<0.31.0

# 任务调度
APScheduler>=3.11.0,<4.0.0

# 环境与配置
python-dotenv>=1.1.0,<2.0.0

# 工具库
PyJWT>=2.10.1,<3.0.0
requests>=2.32.3,<3.0.0

# 生产环境
hypercorn>=0.17.3,<0.18.0

# 加密支持
pycryptodome>=3.22.0,<4.0.0

# 类型提示
pydantic>=2.11.4,<3.0.0
pydantic-settings>=2.9.1,<3.0.0

# 数据保存
PyYAML>=6.0.2,<7.0.0
cryptography>=44.0.3,<45.0.0

# 姓名首字母设置
pypinyin>=0.54.0,<0.55.0

# 其他工具
filelock>=3.18.0,<4.0.0
argon2-cffi>=23.1.0,<24.0.0
redis>=6.0.0,<7.0.0
colorlog>=6.9.0,<7.0.0
